requirements.txt
appdirs==1.4.3
attrs==19.3.0
autobahn==20.1.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.95
cryptography==2.8
dal-admin-filters==0.3.3
daphne==2.4.1
diff-match-patch==20181111
Django==3.0.1
django-autocomplete-light==3.5.0
django-import-export==2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-paypal==1.0.0
django-sendgrid-v5==0.8.0
django-smartfields==1.1.0
django-storages==1.5.2
django-tables2==2.2.1
django-timezone-field==4.0
django-tinymce==2.8.0
docutils==0.16
et-xmlfile==1.0.1
future==0.18.2
gunicorn==19.7.1
importlib-metadata==1.4.0
MarkupPy==1.14
model-mommy==1.6.0
more-itertools==8.1.0
msgpack==0.6.2
odfpy==1.4.0
olefile==0.44
openpyxl==3.0.2
packaging==16.8
pika==0.9.14
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.2.0
python-crontab==2.4.0
python-dateutil==2.6.0
python-decouple==3.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.1.13
sendgrid==6.1.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
Twisted==19.10.0
txaio==20.1.1
unicode-slugify==0.1.3
Unidecode==0.4.20
urllib3==1.25.7
vine==1.3.0
virtualenv==15.1.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
django-bootstrap-datepicker-plus==3.0.5
python-decouple==3.0